moved server side code into lib folder
