+++ lib/cases.spec.ts
