+++ lib/residents.ts
