+++ lib/warningNotes.ts
