+++ pages/api/residents/[id]/cases/[caseId].ts
@@ -1,6 +1,6 @@
 import { StatusCodes } from 'http-status-codes';
 
-import { getCaseByResident } from 'utils/server/cases';
+import { getCaseByResident } from 'lib/cases';
 import { isAuthorised } from 'utils/auth';
 
 import type { NextApiRequest, NextApiResponse, NextApiHandler } from 'next';
