+++ pages/api/teams/[team_id]/workers.ts
@@ -1,6 +1,6 @@
 import { StatusCodes } from 'http-status-codes';
 
-import { getWorkers } from 'utils/server/workers';
+import { getWorkers } from 'lib/workers';
 import { isAuthorised } from 'utils/auth';
 
 import type { NextApiRequest, NextApiResponse, NextApiHandler } from 'next';
