+++ pages/api/teams/index.ts
@@ -1,6 +1,6 @@
 import { StatusCodes } from 'http-status-codes';
 
-import { getTeams } from 'utils/server/teams';
+import { getTeams } from 'lib/teams';
 import { isAuthorised } from 'utils/auth';
 
 import type { NextApiRequest, NextApiResponse, NextApiHandler } from 'next';
